flux/ai/agent-tools/src/requirements.txt
flask>=3.1.0
cloudevents>=1.12.0
kubernetes>=34.1.0
prometheus-client>=0.21.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# === Agent Optimization (Queueing, Game, Control Theory) ===
# Installed from local shared-lib during Docker build
# -e /app/shared-lib
